refactor(perf): rename cubesandbox_clean.py -> business/cleanup.py
